+++ hahomematic/platforms/device.py
@@ -652,16 +652,21 @@ class ValueCache:
         paramset_key: str,
         parameter: str,
         call_source: CallSource,
+        direct_call: bool = False,
     ) -> Any:
         """Load data."""
         async with self._sema_get_or_load_value:
             if (
-                cached_value := self._get_value_from_cache(
-                    channel_address=channel_address,
-                    paramset_key=paramset_key,
-                    parameter=parameter,
+                direct_call is False
+                and (
+                    cached_value := self._get_value_from_cache(
+                        channel_address=channel_address,
+                        paramset_key=paramset_key,
+                        parameter=parameter,
+                    )
                 )
-            ) != NO_CACHE_ENTRY:
+                != NO_CACHE_ENTRY
+            ):
                 return (
                     NO_CACHE_ENTRY if cached_value == self._NO_VALUE_CACHE_ENTRY else cached_value
                 )